Move From impls from hashing to bazel_protos
hashing should be a nice quick/easy/small thing to compile, which
shouldn't rely on protocol buffer compilation because it just has a
couple of value types in.

bazel_protos already needs to depend on protocol buffer compilation.
