+++ src/rust/engine/hashing/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.0.1"
 authors = [ "Pants Build <pantsbuild@gmail.com>" ]
 
 [dependencies]
-bazel_protos = { path = "../process_execution/bazel_protos" }
 digest = "0.6.2"
 hex = "0.3.1"
 sha2 = "0.6.0"
