+++ src/rust/engine/process_execution/bazel_protos/.gitignore
@@ -1,3 +1,4 @@
 src/*.rs
+!src/conversions.rs
 !src/lib.rs
 !src/verification.rs
