+++ src/rust/engine/process_execution/bazel_protos/Cargo.toml
@@ -7,4 +7,5 @@ authors = [ "Pants Build <pantsbuild@gmail.com>" ]
 bytes = "0.4.5"
 futures = "0.1.16"
 grpcio = { version = "0.2.0", features = ["secure"] }
+hashing = { path = "../../hashing" }
 protobuf = { version = "1.4.1", features = ["with-bytes"] }
