+++ src/rust/engine/process_execution/bazel_protos/src/lib.rs
@@ -1,6 +1,7 @@
 extern crate bytes;
 extern crate futures;
 extern crate grpcio;
+extern crate hashing;
 extern crate protobuf;
 
 pub mod bytestream;
@@ -14,5 +15,6 @@ pub mod remote_execution;
 pub mod remote_execution_grpc;
 pub mod status;
 
+mod conversions;
 mod verification;
 pub use verification::verify_directory_canonical;
